Moving test_utils.py into its new home
